tmux/ui: fix window sizes (#2)
This change fixes the window sizes while detached (ie. the preview
window) and attached to conform to the preview pane size and terminal
size respectively.

This change also refactors the *TmuxSession to be more organized and
probably fixes a handful of bugs. Attaching/detaching and goroutines are
managed better now so nothing leaks / is in the wrong state. Things
might get corrupted if an error occurs though. We should address that in
the future.