app: add yolo mode
This change introduces yolo mode. We call this "auto yes" to make it clear that, in this mode, the claude instance automatically hits yes when prompted. This is a new menu
option. If yolo mode is enabled, the instance will have (yolo) next to it's name.

We check for if Yes needs to be tapped at the same time we poll for Running/Ready status. We check if the text "Do you want" is present in the claude output. We could get false positives if the users code has "Do you want" in it, but tapping enter really isn't that bad.

We write if yolo mode is enabled to storage for each instance.

This change also makes sure we truncate titles if the text is too long.