feat(daemon): introduce autoyes daemon
This adds a hidden --daemon flag which reads all the sessions from instances.json and taps yes on them. The daemon process is started by the main process upon exiting. The pid is written to the config directory. On the next startup, the daemon process is looked up and killed by the main process.

The --daemon flag always starts the daemon. It's up to the main process to check if the auto yes flag is enabled before starting the daemon. It always checks if a daemon is runnnig on startup, incase auto yes is disabled but was previously enabled.

Also, our logs were borked bc of the new Info/Error/Warn loggers. Since the functions were in log.go, all the line numbers / file names in the log were in log.go :( To fix this, I updated all the loggers / log calls.