feat(app, session, ui): Add support for poweruser instance creation (#15)
This change adds the ability to submit prompts upon instance creation
(meant for powerusers who want to spin up instances quickly). The
interface can be extended at a later time, if needed, to support sending
messages to the tmux session without having to jump in.

The text input is achieved with an overlay which redraws the background
with a foreground placed on top, we also "fade out" the background to
bring the focus to the text input. Most of the overlay drawing was taken
from https://github.com/mufeez-amjad/lin (past project).


https://github.com/user-attachments/assets/248ff62d-3879-4329-8a7c-904786e76ab7