fix(ui): make menu shorter and improve underlining
This change updates the "pause to checkout" action to just say "checkout". The action key is now "c" instead of "p". This makes the menu shorter.

Also, we now only show shift+up in the diff view. Having both up and down is overkill.

Finally, this changes fixes the `handleKeyPress` control flow when we are not doing menu highlighting. Before, we used to re-send the keypress. This is not efficient. Now, we just handle it in place.